requirements.txt
Flask==1.0.2
MechanicalSoup==0.11.0
google-api-python-client==1.7.8
google-auth==1.6.3
oauth2client==4.1.3
